# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: translator-server API
  version: 1.0.0
  summary: Canonical REST API for the Yara translator-server.
  description: |
    The canonical API surface is mounted under `/api/v1`. Legacy aliases under
    `/api/db/*`, `/api/user/*`, `/api/epubs/*`, `/api/backup/*`,
    `/api/browser-workers`, `/api/proxy/*`, `/api/defaults`,
    `/api/translation-jobs/*` are **deprecated** and will be removed on
    **Wed, 01 Jan 2026** (the `Sunset` header is set on every legacy response).
    New clients must use `/api/v1`.

    When the running server and this document diverge, the running server is the
    source of truth.
  contact:
    name: Yara
    url: https://github.com/mfloresz/yara
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
    description: Canonical v1 surface
  - url: /
    description: Server root (legacy aliases live directly under /api/...)
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /novels/{novelId}/jobs:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    get:
      tags: [Jobs]
      summary: List jobs for a novel.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: failedOnly
          in: query
          schema: { type: integer, enum: [0, 1] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CollectionEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Job' }
    post:
      tags: [Jobs]
      summary: Create a translation / refine / download job.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobCreate' }
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema: { type: string }
              description: '`/api/v1/jobs/<jobId>`'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Job' }
        '503':
          description: Job queue full
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /jobs/{jobId}:
    parameters:
      - $ref: '#/components/parameters/JobIdParam'
    get:
      tags: [Jobs]
      summary: Get one job.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Job' }
        '404': { $ref: '#/components/responses/NotFound' }
  /jobs/{jobId}/cancel:
    parameters:
      - $ref: '#/components/parameters/JobIdParam'
    post:
      tags: [Jobs]
      summary: Cancel a running or pending job.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Job' }
  /jobs/{jobId}/retry:
    parameters:
      - $ref: '#/components/parameters/JobIdParam'
    post:
      tags: [Jobs]
      summary: Re-queue a failed or cancelled job.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Job' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503':
          description: Job queue full
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ---------------------------------------------------------------------------
  # Epubs
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth.token
  parameters:
...
    NovelIdParam:
      name: novelId
      in: path
      required: true
      schema: { type: string }
...
    JobIdParam:
      name: jobId
      in: path
      required: true
      schema: { type: string }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPageParam:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
...
    NotFound:
      description: Not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Conflict.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem details. The envelope uses `error` instead of the standard top-level fields to keep the v1 envelope consistent.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, forbidden, not_found, conflict, queue_full, internal_error]
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
                  code: { type: string }
    SingleEnvelope:
      type: object
      description: Canonical v1 single-resource envelope.
      required: [data]
      properties:
        data: {}
    CollectionEnvelope:
      type: object
      description: Canonical v1 collection envelope.
      required: [data, meta, links]
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/Meta'
        links:
          $ref: '#/components/schemas/Links'
    Meta:
      type: object
      required: [total, has_more]
      properties:
        total: { type: integer, format: int32 }
        page: { type: integer, format: int32 }
        per_page: { type: integer, format: int32 }
        limit: { type: integer, format: int32 }
        offset: { type: integer, format: int32 }
        has_more: { type: boolean }
        next_page: { type: [integer, 'null'], format: int32 }
    Links:
      type: object
      required: [self]
      properties:
        self: { type: string }
        next: { type: string }
        prev: { type: string }
        first: { type: string }
        last: { type: string }
...
    Job:
      type: object
      required: [id, novelId, status, operation]
      properties:
        id: { type: string }
        novelId: { type: string }
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        operation:
          type: string
          enum: [translate, refine, download, check, generate-glossary]
        provider: { type: string }
        model: { type: string }
        totalChapters: { type: integer }
        completedChapters: { type: integer }
        failedChapters: { type: integer }
        errorMessage: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
        novelTitle: { type: string }
        chapterIds:
          type: array
          items: { type: string }
        autoSegmentEnabled: { type: boolean }
        autoSegmentActive: { type: boolean }
        autoSegmentCount: { type: integer }
        autoSegmentCurrentIndex: { type: integer }
        autoSegmentCompletedCount: { type: integer }
        autoSegmentChapterId: { type: string }
        autoSegmentChapterTitle: { type: string }
        newChapters: { type: integer }
    JobCreate:
      type: object
      required: [operation]
      properties:
        operation:
          type: string
          enum: [translate, refine]
        chapterIds:
          type: array
          items: { type: string }
        options:
          type: object
          description: Free-form options; commonly `provider`, `model`, `concurrency`, `autoSegment`, etc.
...